Renamed classifier.js and add comments
